refactor: split SRD into separate components
