+++ client/src/components/SearchableReviewDisplay/components/NAText.tsx
