+++ client/src/components/SearchableReviewDisplay/components/SearchRow.scss
@@ -1,0 +1,7 @@
+.review-reader {
+  .input-search {
+    height: 100%;
+    width: 100%;
+    padding: 24px;
+  } 
+}
