+++ client/src/components/SearchableReviewDisplay/types.ts
@@ -1,0 +1,2 @@
+export type VoidHandler = () => void;
+export type SearchHandler = (q: string) => void;
